fix: decimal compare

+++ database/repository/subscription_repo.go
@@ -46,6 +46,20 @@ func (r *SubscriptionRepository) UpdateSubscription(subscription *model.Subscrip
 	return err
 }
 
+func (r *SubscriptionRepository) GetAllByOwnerId(ownerId primitive.ObjectID) ([]model.Subscription, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), TIMEOUT_DURATION)
+	defer cancel()
+
+	cursor, err := r.SubCol.Find(ctx, bson.M{"ownerId": ownerId})
+	if err != nil {
+		return nil, err
+	}
+
+	var subscriptions []model.Subscription
+	err = cursor.All(ctx, &subscriptions)
+	return subscriptions, err
+}
+
 func (r *SubscriptionRepository) GetAll() ([]model.Subscription, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), TIMEOUT_DURATION)
 	defer cancel()
@@ -60,11 +74,11 @@ func (r *SubscriptionRepository) GetAll() ([]model.Subscription, error) {
 	return subscriptions, err
 }
 
-func (r *SubscriptionRepository) DeleteSubscriptionByName(name string) error {
+func (r *SubscriptionRepository) DeleteSubscriptionByName(name string, ownerId primitive.ObjectID) error {
 	ctx, cancel := context.WithTimeout(context.Background(), TIMEOUT_DURATION)
 	defer cancel()
 
-	result, err := r.SubCol.DeleteOne(ctx, bson.M{"name": name})
+	result, err := r.SubCol.DeleteOne(ctx, bson.M{"name": name, "ownerId": ownerId})
 
 	if err != nil {
 		return err
